Allow field name to differ from the rtl name
The camlp4 extensions allowed

```
module X = interface ABC end
```

The corresponding record would use the field name `abc` but the string
in the `t` value would still be `ABC`.

This is necessary when interfacing with external verilog modules which may
used upper case identifiers, which ocaml wouldn't otherwise allow.

To support this we add a [@rtlname "ABC"] attribute.